feat(ai): export assistant analysis result as markdown
